+++ infra/modules/relayworks-platform/variables.tf
@@ -31,6 +31,7 @@ variable "sync_worker_image" {
 }
 variable "control_plane_api_client_id" { type = string }
 variable "console_client_id" { type = string }
+variable "alert_email" { type = string }
 
 variable "tags" {
   type = map(string)